Add optional data to the success response
The `success` method of the `getJsonHandler` function now accepts an optional second argument, which is an object containing data to be included in the success response. This data will be serialized to JSON and included in the response body.

This change allows for more flexibility in the success response, as it can now include additional information beyond just a success message. For example, it could be used to include the ID of a newly created object, or the results of a query.