"Update validation class and shareApp; remove unused sharedApp functions"
This commit includes updates to the validation class, specifically the removal of the commented out section related to available validators. The shareApp class has also been updated to remove the setExp and getExp methods, as well as the addition of a null check for the appInstance in the getAppInstance method. Lastly, the AppSharedManager has been removed from the Xprz class and is no longer exposed globally.