"Update route handler functions to use context object instead of req and res"
This commit updates the route handler functions to use the `ctx` object instead of the `req` and `res` objects. This change affects the following files:

* `README.md`: The example code in the README file has been updated to use the `ctx` object in the route handler functions.
* `controller/home/home.js`: The `getHome` function has been updated to use the `ctx` object in the route handler function.
* `controller/user/user.js`: The `postSignup` function has been updated to use the `ctx` object in the route handler function.

The purpose of this change is to make the route handler functions more consistent and easier to work with, as the `ctx` object contains both the `req` and `res` objects as properties.