"Add HTTP method prototype methods to RouteManager and remove individual method functions"
This commit adds a new `method()` function to the `RouteManager` class which dynamically creates the prototype methods for each HTTP method (`get`, `post`, `delete`, `put`, `patch`, `options`). It also removes the individual method functions (`get()`, `post()`, etc.) that were previously defined in the class. This reduces redundancy and makes it easier to add or remove HTTP methods in the future. The `method()` function throws an error if an HTTP method is invalid or if a prototype method with the same name already exists.