"Add RouteManager class to manage routes and add routes to Express app"
Explanation:
The code diff shows the addition of a new class, RouteManager, which is used to manage routes and add them to the Express application. The constructor of RouteManager now accepts an optional Express Router object as an argument, with a default value of `getExp().Router()`. This allows for more flexibility in instantiating the RouteManager class. Additionally, the RouteManager class is instantiated and added to the Express app in the `initApp` method of the App class.