"Add verifyBody method to Request, update ReqEnhancer and RouteManager"
This commit adds the `verifyBody` method to the `Request` class, which validates the request body against a set of rules. The `ReqEnhancer` class is updated to remove several utility methods related to query parameters, body parameters, cookies, and headers, as they are now available. The `RouteManager` class is updated to create the request and response objects using the `Request` and `Response` classes, respectively, instead of spreading them with the `req` and `res` objects.